split easydi_sensors into 3 components for systematic naming convention
